Move out clone
